--- a/benchmarks/query-param/src/hono.mts
+++ b/benchmarks/query-param/src/hono.mts
@@ -1,0 +1,5 @@
+import { getQueryParam } from '../../../src/utils/url'
+
+export default (url, key?) => {
+  return getQueryParam(url, key)
+}