--- a/src/utils/url.test.ts
+++ b/src/utils/url.test.ts
@@ -4,7 +4,6 @@ import {
   getPattern,
   getPathFromURL,
   mergePath,
-  getQueryStringFromURL,
   checkOptionalParameter,
   getQueryParam,
   getQueryParams,
@@ -96,23 +95,6 @@ describe('url', () => {
     })
   })
 
-  describe('getQueryStringFromURL', () => {
-    it('should return strings of query params', () => {
-      let queryString = getQueryStringFromURL('https://example.com/?foo=bar')
-      expect(queryString).toBe('foo=bar')
-      queryString = getQueryStringFromURL('https://example.com/?foo=bar&foo2=bar2')
-      expect(queryString).toBe('foo=bar&foo2=bar2')
-      queryString = getQueryStringFromURL('https://example.com/')
-      expect(queryString).toBe('')
-      // This specification allows the fragments as query strings
-      queryString = getQueryStringFromURL('https://example.com/?#foo=#bar&#foo2=#bar2')
-      expect(queryString).toBe('#foo=#bar&#foo2=#bar2')
-      // This specification allows that the string includes two `?` or more
-      queryString = getQueryStringFromURL('https://example.com/?foo=bar?foo2=bar2')
-      expect(queryString).toBe('foo=bar?foo2=bar2')
-    })
-  })
-
   describe('mergePath', () => {
     it('mergePath', () => {
       expect(mergePath('/book', '/')).toBe('/book')
@@ -159,21 +141,21 @@ describe('url', () => {
 
   describe('getQueryParam', () => {
     it('Parse URL query strings', () => {
-      expect(getQueryParam('name=hey', 'name')).toBe('hey')
-      expect(getQueryParam('name=hey#fragment', 'name')).toBe('hey#fragment')
-      expect(getQueryParam('name=hey&age=20&tall=170', 'age')).toBe('20')
+      expect(getQueryParam('http://example.com/?name=hey', 'name')).toBe('hey')
+      expect(getQueryParam('http://example.com/?name=hey#fragment', 'name')).toBe('hey#fragment')
+      expect(getQueryParam('http://example.com/?name=hey&age=20&tall=170', 'age')).toBe('20')
       let searchParams = new URLSearchParams({ name: '炎' })
-      expect(getQueryParam(searchParams.toString(), 'name')).toBe('炎')
-      expect(getQueryParam('name=hey&age=20&tall=170', 'weight')).toBe(null)
-      expect(getQueryParam('name=hey&age=20&tall=170')).toEqual({
+      expect(getQueryParam(`http://example.com/?${searchParams.toString()}`, 'name')).toBe('炎')
+      expect(getQueryParam('http://example.com/?name=hey&age=20&tall=170', 'weight')).toBe(undefined)
+      expect(getQueryParam('http://example.com/?name=hey&age=20&tall=170')).toEqual({
         name: 'hey',
         age: '20',
         tall: '170',
       })
-      expect(getQueryParam('pretty', 'pretty')).toBe('')
-      expect(getQueryParam('pretty', 'prtt')).toBe(null)
-      expect(getQueryParam('name=sam&name=tom', 'name')).toBe('sam')
-      expect(getQueryParam('name=sam&name=tom')).toEqual({
+      expect(getQueryParam('http://example.com/?pretty', 'pretty')).toBe('')
+      expect(getQueryParam('http://example.com/?pretty', 'prtt')).toBe(undefined)
+      expect(getQueryParam('http://example.com/?name=sam&name=tom', 'name')).toBe('sam')
+      expect(getQueryParam('http://example.com/?name=sam&name=tom')).toEqual({
         name: 'sam',
       })
       searchParams = new URLSearchParams('?name=sam=tom')
@@ -183,26 +165,26 @@ describe('url', () => {
 
   describe('getQueryParams', () => {
     it('Parse URL query strings', () => {
-      expect(getQueryParams('name=hey', 'name')).toEqual(['hey'])
-      expect(getQueryParams('name=hey#fragment', 'name')).toEqual(['hey#fragment'])
-      expect(getQueryParams('name=hey&name=foo', 'name')).toEqual(['hey', 'foo'])
-      expect(getQueryParams('name=hey&age=20&tall=170', 'age')).toEqual(['20'])
-      expect(getQueryParams('name=hey&age=20&tall=170&name=foo&age=30', 'age')).toEqual([
+      expect(getQueryParams('http://example.com/?name=hey', 'name')).toEqual(['hey'])
+      expect(getQueryParams('http://example.com/?name=hey#fragment', 'name')).toEqual(['hey#fragment'])
+      expect(getQueryParams('http://example.com/?name=hey&name=foo', 'name')).toEqual(['hey', 'foo'])
+      expect(getQueryParams('http://example.com/?name=hey&age=20&tall=170', 'age')).toEqual(['20'])
+      expect(getQueryParams('http://example.com/?name=hey&age=20&tall=170&name=foo&age=30', 'age')).toEqual([
         '20',
         '30',
       ])
       const searchParams = new URLSearchParams()
       searchParams.append('tag', '炎')
       searchParams.append('tag', 'ほのお')
-      expect(getQueryParams(searchParams.toString(), 'tag')).toEqual(['炎', 'ほのお'])
-      expect(getQueryParams('name=hey&age=20&tall=170', 'weight')).toEqual(null)
-      expect(getQueryParams('name=hey&age=20&tall=170&name=foo&age=30&tall=180')).toEqual({
+      expect(getQueryParams(`http://example.com/?${searchParams.toString()}`, 'tag')).toEqual(['炎', 'ほのお'])
+      expect(getQueryParams('http://example.com/?name=hey&age=20&tall=170', 'weight')).toEqual(undefined)
+      expect(getQueryParams('http://example.com/?name=hey&age=20&tall=170&name=foo&age=30&tall=180')).toEqual({
         name: ['hey', 'foo'],
         age: ['20', '30'],
         tall: ['170', '180'],
       })
-      expect(getQueryParams('pretty', 'pretty')).toEqual([''])
-      expect(getQueryParams('pretty', 'prtt')).toBe(null)
+      expect(getQueryParams('http://example.com/?pretty', 'pretty')).toEqual([''])
+      expect(getQueryParams('http://example.com/?pretty', 'prtt')).toBe(undefined)
     })
   })
 })