More accurate decoding in query string parsing. (e.g., encoded keys)

### performance

The following are the results of a comparison between hono's v3.1.3 and fast-querystring (which is used in elysia) with a slight modification to the benchmarks/query-param added in this PR.

In some places, hono's v3.1.3 is faster because decoding is skipped, and compared to that, this PR has a performance decrease. Nevertheless, the speed reduction is slight, and the amount of code has not increased.
This should be changed as it is unsuitable for a framework not to get parameters right.

#### bun

```
$ npm run bench:bun

> bench:bun
> bun run ./src/bench.mts

cpu: Apple M2 Pro
runtime: bun 0.5.8 (arm64-darwin)

benchmark            time (avg)             (min … max)       p75       p99      p995
------------------------------------------------------- -----------------------------
• {"url":"http://example.com/?page=1","key":"page"}
------------------------------------------------------- -----------------------------
hono 3.1.3        65.82 ns/iter  (61.54 ns … 198.54 ns)  65.23 ns 117.97 ns 133.18 ns
hono with PR      66.68 ns/iter  (60.56 ns … 174.03 ns)  66.41 ns 123.95 ns  138.5 ns
fastQuerystring   90.34 ns/iter  (84.92 ns … 269.87 ns)  88.87 ns 143.85 ns 167.52 ns

summary for {"url":"http://example.com/?page=1","key":"page"}
  hono 3.1.3
   1.01x faster than hono with PR
   1.37x faster than fastQuerystring

• {"url":"http://example.com/?url=http://example.com&page=1","key":"page"}
------------------------------------------------------- -----------------------------
hono 3.1.3       158.91 ns/iter  (150.2 ns … 333.56 ns) 155.87 ns 217.88 ns 324.67 ns
hono with PR     121.36 ns/iter (111.56 ns … 219.51 ns) 119.86 ns 184.04 ns    193 ns
fastQuerystring  161.95 ns/iter  (152.05 ns … 230.4 ns) 163.67 ns 225.31 ns 227.72 ns

summary for {"url":"http://example.com/?url=http://example.com&page=1","key":"page"}
  hono with PR
   1.31x faster than hono 3.1.3
   1.33x faster than fastQuerystring

• {"url":"http://example.com/?page=1"}
------------------------------------------------------- -----------------------------
hono 3.1.3        93.51 ns/iter   (88.1 ns … 216.14 ns)  92.46 ns 146.49 ns 192.24 ns
hono with PR      98.15 ns/iter   (93.23 ns … 176.8 ns)  97.18 ns 149.85 ns 162.26 ns
fastQuerystring   88.39 ns/iter  (82.92 ns … 211.82 ns)  87.12 ns 143.22 ns 178.06 ns

summary for {"url":"http://example.com/?page=1"}
  fastQuerystring
   1.06x faster than hono 3.1.3
   1.11x faster than hono with PR

• {"url":"http://example.com/?url=http://example.com&page=1"}
------------------------------------------------------- -----------------------------
hono 3.1.3       181.37 ns/iter (170.66 ns … 302.08 ns) 180.01 ns 243.45 ns 247.03 ns
hono with PR     186.77 ns/iter (176.45 ns … 290.13 ns)  190.2 ns  246.7 ns 263.16 ns
fastQuerystring  162.25 ns/iter  (153.31 ns … 302.2 ns) 159.61 ns 227.72 ns 247.82 ns

summary for {"url":"http://example.com/?url=http://example.com&page=1"}
  fastQuerystring
   1.12x faster than hono 3.1.3
   1.15x faster than hono with PR

• {"url":"http://example.com/?url=http://example.com/very/very/deep/path/to/something&search=very-long-search-string"}
------------------------------------------------------- -----------------------------
hono 3.1.3       196.09 ns/iter  (183.96 ns … 276.2 ns) 194.43 ns 257.71 ns 272.48 ns
hono with PR     232.93 ns/iter (222.82 ns … 309.85 ns) 231.97 ns 294.33 ns 299.66 ns
fastQuerystring  259.47 ns/iter (249.71 ns … 319.48 ns) 258.02 ns 315.05 ns 315.56 ns

summary for {"url":"http://example.com/?url=http://example.com/very/very/deep/path/to/something&search=very-long-search-string"}
  hono 3.1.3
   1.19x faster than hono with PR
   1.32x faster than fastQuerystring

• {"url":"http://example.com/?search=Hono+is+a+small,+simple,+and+ultrafast+web+framework+for+the+Edge.&page=1"}
------------------------------------------------------- -----------------------------
hono 3.1.3       184.92 ns/iter (174.18 ns … 326.26 ns) 187.11 ns 254.86 ns 256.74 ns
hono with PR     621.62 ns/iter (576.61 ns … 810.12 ns)  623.6 ns 810.12 ns 810.12 ns
fastQuerystring  542.04 ns/iter (516.23 ns … 739.35 ns) 538.62 ns 710.87 ns 739.35 ns

summary for {"url":"http://example.com/?search=Hono+is+a+small,+simple,+and+ultrafast+web+framework+for+the+Edge.&page=1"}
  hono 3.1.3
   2.93x faster than fastQuerystring
   3.36x faster than hono with PR

• {"url":"http://example.com/?a=1&b=2&c=3&d=4&e=5&f=6&g=7&h=8&i=9&j=10"}
------------------------------------------------------- -----------------------------
hono 3.1.3       628.45 ns/iter (598.59 ns … 732.02 ns) 638.18 ns 732.02 ns 732.02 ns
hono with PR     658.52 ns/iter  (525.1 ns … 733.42 ns) 661.84 ns 733.42 ns 733.42 ns
fastQuerystring   301.3 ns/iter (283.41 ns … 388.77 ns) 302.77 ns 383.02 ns 388.77 ns

summary for {"url":"http://example.com/?a=1&b=2&c=3&d=4&e=5&f=6&g=7&h=8&i=9&j=10"}
  fastQuerystring
   2.09x faster than hono 3.1.3
   2.19x faster than hono with PR
```

#### node

```
$ npm run bench:node

> bench:node
> tsx ./src/bench.mts

cpu: Apple M2 Pro
runtime: node v18.14.0 (arm64-darwin)

benchmark            time (avg)             (min … max)       p75       p99      p995
------------------------------------------------------- -----------------------------
• {"url":"http://example.com/?page=1","key":"page"}
------------------------------------------------------- -----------------------------
hono 3.1.3        41.39 ns/iter  (37.95 ns … 516.84 ns)  43.52 ns  57.09 ns  60.73 ns
hono with PR      55.12 ns/iter     (52.5 ns … 68.3 ns)  57.14 ns  62.19 ns  64.97 ns
fastQuerystring   66.02 ns/iter  (63.96 ns … 372.19 ns)   65.9 ns  72.42 ns  73.45 ns

summary for {"url":"http://example.com/?page=1","key":"page"}
  hono 3.1.3
   1.33x faster than hono with PR
   1.6x faster than fastQuerystring

• {"url":"http://example.com/?url=http://example.com&page=1","key":"page"}
------------------------------------------------------- -----------------------------
hono 3.1.3        124.2 ns/iter  (118.8 ns … 131.31 ns) 126.15 ns 130.28 ns 130.52 ns
hono with PR      76.68 ns/iter    (73.77 ns … 88.6 ns)  79.09 ns  83.51 ns  84.45 ns
fastQuerystring  199.72 ns/iter (194.35 ns … 217.53 ns) 202.18 ns 211.56 ns 211.94 ns

summary for {"url":"http://example.com/?url=http://example.com&page=1","key":"page"}
  hono with PR
   1.62x faster than hono 3.1.3
   2.6x faster than fastQuerystring

• {"url":"http://example.com/?page=1"}
------------------------------------------------------- -----------------------------
hono 3.1.3        80.57 ns/iter  (77.79 ns … 104.94 ns)  82.96 ns  89.03 ns  99.25 ns
hono with PR      91.85 ns/iter    (87.9 ns … 107.6 ns)   93.9 ns 101.43 ns 103.75 ns
fastQuerystring   71.55 ns/iter   (68.54 ns … 86.92 ns)  74.12 ns  82.68 ns  82.95 ns

summary for {"url":"http://example.com/?page=1"}
  fastQuerystring
   1.13x faster than hono 3.1.3
   1.28x faster than hono with PR

• {"url":"http://example.com/?url=http://example.com&page=1"}
------------------------------------------------------- -----------------------------
hono 3.1.3       164.39 ns/iter (153.94 ns … 197.98 ns) 168.11 ns 186.43 ns 196.59 ns
hono with PR     167.32 ns/iter (163.34 ns … 181.16 ns) 169.12 ns 176.17 ns 177.04 ns
fastQuerystring  202.08 ns/iter (195.31 ns … 213.89 ns) 204.75 ns  213.7 ns  213.7 ns

summary for {"url":"http://example.com/?url=http://example.com&page=1"}
  hono 3.1.3
   1.02x faster than hono with PR
   1.23x faster than fastQuerystring

• {"url":"http://example.com/?url=http://example.com/very/very/deep/path/to/something&search=very-long-search-string"}
------------------------------------------------------- -----------------------------
hono 3.1.3       161.58 ns/iter (156.25 ns … 172.65 ns)  163.5 ns 168.73 ns 172.38 ns
hono with PR     200.47 ns/iter (195.61 ns … 212.69 ns) 202.35 ns 210.52 ns 211.87 ns
fastQuerystring  400.78 ns/iter (393.91 ns … 439.39 ns) 402.97 ns 426.57 ns 439.39 ns

summary for {"url":"http://example.com/?url=http://example.com/very/very/deep/path/to/something&search=very-long-search-string"}
  hono 3.1.3
   1.24x faster than hono with PR
   2.48x faster than fastQuerystring

• {"url":"http://example.com/?search=Hono+is+a+small,+simple,+and+ultrafast+web+framework+for+the+Edge.&page=1"}
------------------------------------------------------- -----------------------------
hono 3.1.3       165.62 ns/iter (159.12 ns … 206.62 ns) 167.27 ns 182.45 ns 203.05 ns
hono with PR        646 ns/iter (634.85 ns … 691.84 ns) 647.97 ns 691.84 ns 691.84 ns
fastQuerystring  724.25 ns/iter (711.73 ns … 763.09 ns) 725.07 ns 763.09 ns 763.09 ns

summary for {"url":"http://example.com/?search=Hono+is+a+small,+simple,+and+ultrafast+web+framework+for+the+Edge.&page=1"}
  hono 3.1.3
   3.9x faster than hono with PR
   4.37x faster than fastQuerystring

• {"url":"http://example.com/?a=1&b=2&c=3&d=4&e=5&f=6&g=7&h=8&i=9&j=10"}
------------------------------------------------------- -----------------------------
hono 3.1.3       534.15 ns/iter  (524.4 ns … 561.31 ns) 536.08 ns 552.36 ns 561.31 ns
hono with PR     419.28 ns/iter (410.06 ns … 449.23 ns) 421.47 ns 439.72 ns 449.23 ns
fastQuerystring  397.81 ns/iter (388.85 ns … 414.47 ns) 400.37 ns 412.43 ns 414.47 ns

summary for {"url":"http://example.com/?a=1&b=2&c=3&d=4&e=5&f=6&g=7&h=8&i=9&j=10"}
  fastQuerystring
   1.05x faster than hono with PR
   1.34x faster than hono 3.1.3
```