@yusukebe 2023-03-29
Hi @usualoma !

Thanks for the PR.

I've implemented the decoding query parameters, but I needed more confidence. So I am thankful that this PR supports decoding `+` and keys with spaces. And I think it is good that you have created a benchmark. As for performance, it is faster than `fast-querystring`, which is great.

I'll merge it!