@tangye1234 2023-03-31
This change brings new bug.

`?pretty&foo=bar` could be resolved with `{ foo: 'bar' }`, but now cannot be resolved